requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
channels==4.0.0
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
outcome==1.2.0
pkg_resources==0.0.0
pyperclip==1.8.2
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
